Always sanitize snapshot file name
